refactor: Move PropertyPathFilterDefinition to PropertyPathFilter
