+++ library/ExternalContent/PropertyPathFilter/FilterDefinition/Contracts/Enums/Operator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\Enums;
+
+enum Operator
+{
+    case EQUALS;
+    case NOT_EQUALS;
+}
