+++ library/ExternalContent/PropertyPathFilter/FilterDefinition/Contracts/Enums/Relation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\Enums;
+
+enum Relation
+{
+    case AND;
+    case OR;
+}
