+++ library/ExternalContent/PropertyPathFilter/FilterDefinition/Contracts/RuleSetInterface.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts;
+namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts;
 
-use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\Enums\Relation;
 
 interface RuleSetInterface
 {
