+++ library/ExternalContent/PropertyPathFilter/FilterDefinition/Filter.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilterDefinition;
+namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition;
 
-use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\FilterInterface;
-use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\RuleSetInterface;
+use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\FilterInterface;
+use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\RuleSetInterface;
 
 /**
  * Class Filter
