+++ library/ExternalContent/PropertyPathFilter/FilterDefinition/RuleSet.test.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilterDefinition;
+namespace Municipio\ExternalContent\PropertyPathFilter\FilterDefinition;
 
-use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Relation;
-use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\RuleInterface;
+use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilter\FilterDefinition\Contracts\RuleInterface;
 use PHPUnit\Framework\TestCase;
 
 class RuleSetTest extends TestCase
