--- library/ExternalContent/PropertyPathFilterDefinition/Contracts/Enums/Operator.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums;
-
-enum Operator
-{
-    case EQUALS;
-    case NOT_EQUALS;
-}
